+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 - Handled inconsistent cross-platform firing of `compositionend` events
 - Accurately calculate caret position for multi-codepoint emojis.
 - #659: Prevent re-opening of dialog when closing with keyboard.
+- Fixed selected output functionality in tutorial.
 
 ## 0.14.2 2024-01-18